Refactored code for readability and user interaction enhancement
This commit includes several changes aimed at improving the structure of the code as well as enhancing user interaction. Classes were either renamed or transferred to relevant categories for clear distinction between import and export operations. The ConsoleReport and CsvReport became ExportConsoleReport and ExportCsvReport as part of the Export category. The new ImportCsvTransactions class was also added to handle csv transactions importation.

Additionally, the user interface was improved, requiring users to input the path to the transaction file. This allows more flexibility, as the file does not have to be in a fixed location anymore. This update also includes instructions added to the README to guide users on how to use the application.